requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
aiohttp==3.9.1
sqlmodel==0.0.14
python-dotenv==1.0.0
loguru==0.7.2
validators==0.22.0
dnspython==2.4.2

# Optional dependencies (will gracefully fallback if not available)
geoip2==4.7.0
maxminddb==2.2.0
user-agents==2.2.0